fixes panel and filter
