+++ bims/api_views/category_filter.py
@@ -5,13 +5,12 @@ from rest_framework.views import APIView
 from bims.models.biological_collection_record import BiologicalCollectionRecord
 
 
-class DateCategoryList(APIView):
-    """API for listing all biological collection record collectors."""
+class CategoryList(APIView):
+    """API for listing all biological collection record category."""
 
     def get(self, request, *args):
         category = \
             BiologicalCollectionRecord.objects.all().values_list(
                 'category', flat=True).distinct().order_by('category')
-        print(category)
         return HttpResponse(
             json.dumps(list(category)), content_type='application/json')
