+++ bims/static/js/router.js
@@ -7,26 +7,6 @@ define(['backbone', 'views/olmap', 'shared'], function (Backbone, olmap, Shared)
         },
         initialize: function () {
             this.map = new olmap();
-            $.ajax({
-                type: 'GET',
-                url: listCollectorAPIUrl,
-                dataType: 'json',
-                success: function (data) {
-                    for (var i = 0; i < data.length; i++) {
-                        $('#filter-collectors').append('<input type="checkbox" name="collector-value" value="' + data[i] + '"> ' + data[i] + '<br>');
-                    }
-                }
-            });
-            $.ajax({
-                type: 'GET',
-                url: listCategoryAPIUrl,
-                dataType: 'json',
-                success: function (data) {
-                    for (var i = 0; i < data.length; i++) {
-                        $('#filter-category').append('<input type="checkbox" name="category-value" value="' + data[i] + '">&nbsp;' + data[i] + '<br>');
-                    }
-                }
-            });
         },
         search: function (query) {
             Shared.Dispatcher.trigger('search:searchCollection', query);
