+++ bims/templates/map.html
@@ -128,7 +128,7 @@
                     <table id="month-selector" style="width:100%">
                     </table>
                     <button class="apply-filter">Apply Filter</button>
-                    <button class="apply-filter">Clear Filter</button>
+                    <button class="clear-filter">Clear Filter</button>
                 </div>
             </div>
             <div class="row sub-filter">
@@ -141,7 +141,7 @@
                 </div>
                 <div class="button-filter-wrapper">
                     <button class="apply-filter">Apply Filter</button>
-                    <button class="apply-filter">Clear Filter</button>
+                    <button class="clear-filter">Clear Filter</button>
                 </div>
             </div>
             <div class="row sub-filter">
@@ -154,7 +154,7 @@
                 </div>
                 <div class="button-filter-wrapper">
                     <button class="apply-filter">Apply Filter</button>
-                    <button class="apply-filter">Clear Filter</button>
+                    <button class="clear-filter">Clear Filter</button>
                 </div>
             </div>
         </div>
