+++ bims/urls.py
@@ -26,7 +26,7 @@ from bims.api_views.cluster_collection_by_taxon import (
     ClusterCollectionByTaxonExtent
 )
 from bims.api_views.collector import CollectorList
-from bims.api_views.date_category_filter import DateCategoryList
+from bims.api_views.category_filter import CategoryList
 from bims.views.bibliography import (
     EntryListView,
     EntryBatchImportView,
@@ -62,8 +62,8 @@ api_urls = [
         SearchObjects.as_view(), name='search-api'),
     url(r'^api/list-collector/$',
         CollectorList.as_view(), name='list-collector'),
-    url(r'^api/list-date-category/$',
-        DateCategoryList.as_view(), name='list-date-category'),
+    url(r'^api/list-category/$',
+        CategoryList.as_view(), name='list-date-category'),
 ]
 
 bibliography_urls = [
